+++ src/Rulesage.Shared/Repositories/Abstractions/IRuleRepository.cs
@@ -6,6 +6,6 @@ public interface IRuleRepository : IDocumentRepository
 {
     Task<RuleExpr?> FindByIdAsync(string id, CancellationToken cancellationToken = default);
 
-    Task<IEnumerable<(RuleExpr, float)>> FindOrderByCosineDistanceAsync(double[] queryVector, int skip, int take,
+    Task<IEnumerable<(RuleExpr, float)>> FindOrderByCosineDistanceAsync(float[] queryVector, int skip, int take,
         CancellationToken cancellationToken = default);
 }
