+++ src/Rulesage.Shared/Services/Abstractions/IEmbeddingService.cs
@@ -5,6 +5,6 @@ public interface IEmbeddingService
     protected const int MaxSequenceLength = 256;
     protected const int OverlapSize = 50;
     
-    double[] GetEmbedding(string text, int chunkSize = MaxSequenceLength, int overlapSize = OverlapSize);
-    double[][] GetBatchEmbeddings(IEnumerable<string> texts, int chunkSize = MaxSequenceLength, int overlapSize = OverlapSize);
+    float[] GetEmbedding(string text, int chunkSize = MaxSequenceLength, int overlapSize = OverlapSize);
+    float[][] GetBatchEmbeddings(IEnumerable<string> texts, int chunkSize = MaxSequenceLength, int overlapSize = OverlapSize);
 }
